Rename `constanats` to `constants` as intended
Fixes #98.
